cic(#1513): lift the joined-channel projection out of the composer closure
`joinedChannelsOnNetwork` is what still pinned `ame`/`amsg` to
`compose.ts` after the pipeline moved: it sat INSIDE the
`identityScopedStore` closure and was never re-exported, so no command
handler could reach it.

It must NOT be exported from `compose.ts` — `compose` imports
`commands/*`, so a handler importing it back would close exactly the
cycle this bucket exists to remove. That constraint is what forces it out
of the file; where it lands was decided by a measurement.

The lift itself was gated on one. Every token of the body was enumerated
and classified — parameter, locals, member accesses, literals, the
`Object` global — leaving two free identifiers, `decodeChannelKey` and
`windowStateByChannel`, both module-scope imports with no shadowing
declaration anywhere in `compose.ts`. Closure captures: zero. Had it
touched one variable from inside the closure this would have been a store
change, not a move, and would have stopped.

The planned destination was `windowState.ts`, whose store it projects.
The gate refused it: `compose.test.ts` replaces that module wholesale
with a `vi.mock` of stubs and drives the tests by setting
`windowStateByChannel`'s return, so a projection placed inside it is
swallowed by the mock — 22 reds across the #30 tab-complete and #431
fan-out sets. Both ways of keeping it there are worse than moving it:
writing the projection into the mock re-implements production logic and
turns those 22 into assertions about the stub, and un-mocking the store
rewrites the CP17 setup they depend on. So it lives in
`joinedChannels.ts`, which reads `windowStateByChannel` from OUTSIDE —
the mock still governs it and the 22 kept exercising the real projection,
with no test edited.

The general rule, recorded because it will recur: a module that consumers
mock wholesale as a state SOURCE is not a home for derived read-model
logic. Put the derivation in a module that READS the store, not in the
one that IS the store.

The DESIGN_NOTES entry rides with this commit, separator included, and
records the three things that will not be re-derivable later: why the
cluster travelled whole rather than as two helpers (all three call sites
use `draftLines`/`wireBody` as a PAIR), the #1192 precedent written into
`ctcpAction.ts:23-26`, and the failure mode the mutant intercepts — a
DUPLICATED rather than moved helper, leaving the wire path and #1108's
preview silently disagreeing, which no passing suite reports. It also
records the correction that "unblocks three of six" was false:
`service-modal` is the only arm clean on the hoist alone.

#1513 stays open. `privmsg`, `me` and `msg` reach `sendPacedBody`, whose
four store verbs DO close over composer state — that half means moving
the store, not the pipeline.
